Update mortgage_e2e_0.8_ETL.ipynb
Updated with improved cudf etl, but still gets stuck on a merge, as noted in the comments.  Probably runs out of memory, so we're movng to Dask Cudf.  Handing back to @DarrenR96.  Once the dask `drop` error is overcome, dask cudf, we should see if the memory management gets it through the merge