21.10 Notebook Testing Report
Discovered errors in a few notebooks in the notebooks-contrib repo after running them in the 21.10 stable release. The errors were tested and found in both CentOS and Ubuntu operating systems. There is a issue for the notebook not yet resolved The current status of all these notebooks with and without a proposed resolution are below:

- [ ] Hello_Streamz
- [x] Federated_query_demo
- [x] 06_Introduction_to_Supervised_Learning
- [x] linear_regression_demo
- [x] rf_demo
- [x] umap_demo_full